Server/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pymongo==4.2.0